Release 5.5
core: Bugfix in warm to initialize the interpreter for turnkey. Thanks to David.
core: bugfixes for handling some negative numbers in +loop and */.
core: simplified assembly primitves for counted loops. They are now 
   faster except for i. The return stack gets different numbers now.
core: rewrite of accept. The user visible change is that the final 
  CR/LF is no longer sent here. The forth text interpreter does it 
  elsewhere thus the user interface is unchanged.
core: Fixed a regression introduced after 5.1: a<b is not always the same as a-b<0.